- Init navigation controller
